pyproject.toml
[project]
name = "finhub"
version = "0.1.0"
description = ""
authors = [
    {name = "jm",email = "e.agafonov@itgene.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (==4.2)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "django-redis (>=5.4.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
